+++ src/iexec-orderbook.js
@@ -9,12 +9,14 @@ const {
   desc,
   option,
   Spinner,
+  displayPaginableRequest,
   pretty,
   info,
   getPropertyFormChain,
 } = require('./cli-helper');
 const { loadChain } = require('./chains');
 const orderbook = require('./orderbook');
+const { NULL_BYTES32, NULL_ADDRESS } = require('./utils');
 
 const objName = 'orderbook';
 
@@ -47,9 +49,10 @@ orderbookApp
         requireTag,
         maxTag,
         minVolume,
+        raw,
       } = cmd;
-      spinner.start(info.showing(objName));
-      const response = await orderbook.fetchAppOrderbook(
+
+      const request = orderbook.fetchAppOrderbook(
         chain.contracts,
         getPropertyFormChain(chain, 'iexecGateway'),
         app,
@@ -62,23 +65,57 @@ orderbookApp
           minVolume,
         },
       );
-      const appOrders = response.appOrders
-        ? response.appOrders.map(e => ({
+      const fetchMessage = info.showing(objName);
+      const processResponse = res => (res.orders
+        ? res.orders.map(e => ({
           orderHash: e.orderHash,
-          app: e.order.app,
-          tag: e.order.tag,
           price: e.order.appprice,
           remaining: e.remaining,
+          app: e.order.app,
+          tag: e.order.tag,
+          datasetrestrict: e.order.datasetrestrict,
+          workerpoolrestrict: e.order.workerpoolrestrict,
+          requesterrestrict: e.order.requesterrestrict,
         }))
-        : [];
+        : []);
+      const createResultsMessage = (
+        callResults,
+        initilResultsCount,
+        totalCount,
+      ) => `Apporders details (${initilResultsCount + 1} to ${initilResultsCount
+          + callResults.length}${totalCount ? ` of ${totalCount}` : ''}):${pretty(
+        callResults.map(e => ({
+          orderHash: e.orderHash,
+          price: e.price,
+          remaining: e.remaining,
+          ...(e.tag !== NULL_BYTES32 && { tag: e.tag }),
+          ...(e.datasetrestrict !== NULL_ADDRESS && {
+            datasetrestrict: e.datasetrestrict,
+          }),
+          ...(e.workerpoolrestrict !== NULL_ADDRESS && {
+            workerpoolrestrict: e.workerpoolrestrict,
+          }),
+          ...(e.requesterrestrict !== NULL_ADDRESS && {
+            requesterrestrict: e.requesterrestrict,
+          }),
+        })),
+      )}`;
 
-      const successMessage = appOrders.length > 0
-        ? `Orderbook\nApp orders details:${pretty(appOrders)}\n`
-        : 'Empty orderbook';
+      const { results, count } = await displayPaginableRequest({
+        request,
+        processResponse,
+        fetchMessage,
+        createResultsMessage,
+        spinner,
+        raw,
+      });
+
+      const successMessage = results.length > 0 ? 'No more results' : 'Empty orderbook';
 
       spinner.succeed(successMessage, {
         raw: {
-          appOrders: response.appOrders,
+          count,
+          appOrders: results,
         },
       });
       spinner.info('Trade in the browser at https://market.iex.ec');
@@ -107,11 +144,17 @@ orderbookDataset
         spinner,
       });
       const {
-        app, workerpool, requester, minVolume,
+        app,
+        workerpool,
+        requester,
+        minVolume,
+        tag,
+        maxTag,
+        requireTag,
+        raw,
       } = cmd;
-      spinner.start(info.showing(objName));
-      const { tag, maxTag, requireTag } = cmd;
-      const response = await orderbook.fetchDatasetOrderbook(
+
+      const request = orderbook.fetchDatasetOrderbook(
         chain.contracts,
         getPropertyFormChain(chain, 'iexecGateway'),
         dataset,
@@ -124,24 +167,59 @@ orderbookDataset
           minVolume,
         },
       );
-      const datasetOrders = response.datasetOrders
-        ? response.datasetOrders.map(e => ({
+      const fetchMessage = info.showing(objName);
+      const processResponse = res => (res.orders
+        ? res.orders.map(e => ({
           orderHash: e.orderHash,
+          price: e.order.datasetprice,
+          remaining: e.remaining,
           dataset: e.order.dataset,
           tag: e.order.tag,
           apprestrict: e.order.apprestrict,
-          price: e.order.datasetprice,
-          remaining: e.remaining,
+          workerpoolrestrict: e.order.workerpoolrestrict,
+          requesterrestrict: e.order.requesterrestrict,
         }))
-        : [];
+        : []);
+      const createResultsMessage = (
+        callResults,
+        initilResultsCount,
+        totalCount,
+      ) => `Datasetorders details (${initilResultsCount
+          + 1} to ${initilResultsCount + callResults.length}${
+        totalCount ? ` of ${totalCount}` : ''
+      }):${pretty(
+        callResults.map(e => ({
+          orderHash: e.orderHash,
+          price: e.price,
+          remaining: e.remaining,
+          ...(e.tag !== NULL_BYTES32 && { tag: e.tag }),
+          ...(e.apprestrict !== NULL_ADDRESS && {
+            apprestrict: e.apprestrict,
+          }),
+          ...(e.workerpoolrestrict !== NULL_ADDRESS && {
+            workerpoolrestrict: e.workerpoolrestrict,
+          }),
+          ...(e.requesterrestrict !== NULL_ADDRESS && {
+            requesterrestrict: e.requesterrestrict,
+          }),
+        })),
+      )}`;
 
-      const successMessage = datasetOrders.length > 0
-        ? `Orderbook\nDataset orders details:${pretty(datasetOrders)}\n`
-        : 'Empty orderbook';
+      const { results, count } = await displayPaginableRequest({
+        request,
+        processResponse,
+        fetchMessage,
+        createResultsMessage,
+        spinner,
+        raw,
+      });
+
+      const successMessage = results.length > 0 ? 'No more results' : 'Empty orderbook';
 
       spinner.succeed(successMessage, {
         raw: {
-          datasetOrders: response.datasetOrders,
+          count,
+          datasetOrders: results,
         },
       });
       spinner.info('Trade in the browser at https://market.iex.ec');
@@ -164,14 +242,13 @@ orderbookWorkerpool
   .option(...option.includeDatasetSpecific())
   .option(...option.includeRequesterSpecific())
   .description(desc.showObj('workerpools orderbook', 'marketplace'))
-  .action(async (address, cmd) => {
+  .action(async (workerpool, cmd) => {
     await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const chain = await loadChain(cmd.chain, {
         spinner,
       });
-      spinner.start(info.showing(objName));
       const {
         tag,
         maxTag,
@@ -182,13 +259,15 @@ orderbookWorkerpool
         category,
         minVolume,
         minTrust,
+        raw,
       } = cmd;
-      const response = await orderbook.fetchWorkerpoolOrderbook(
+
+      const request = orderbook.fetchWorkerpoolOrderbook(
         chain.contracts,
         getPropertyFormChain(chain, 'iexecGateway'),
         {
           category,
-          workerpool: address,
+          workerpool,
           minTag: tag !== undefined ? tag : requireTag,
           maxTag: tag !== undefined ? tag : maxTag,
           minVolume,
@@ -198,25 +277,66 @@ orderbookWorkerpool
           requester,
         },
       );
-      const workerpoolOrders = response.workerpoolOrders
-        ? response.workerpoolOrders.map(e => ({
+      const fetchMessage = info.showing(objName);
+      const processResponse = res => (res.orders
+        ? res.orders.map(e => ({
           orderHash: e.orderHash,
-          workerpool: e.order.workerpool,
+          price: e.order.workerpoolprice,
+          remaining: e.remaining,
           category: e.order.category,
           tag: e.order.tag,
           trust: e.order.trust,
-          price: e.order.workerpoolprice,
-          remaining: e.remaining,
+          workerpool: e.order.workerpool,
+          apprestrict: e.order.apprestrict,
+          datasetrestrict: e.order.datasetrestrict,
+          requesterrestrict: e.order.requesterrestrict,
         }))
-        : [];
+        : []);
+      const createResultsMessage = (
+        callResults,
+        initilResultsCount,
+        totalCount,
+      ) => `Workerpoolorders details (${initilResultsCount
+          + 1} to ${initilResultsCount + callResults.length}${
+        totalCount ? ` of ${totalCount}` : ''
+      }):${pretty(
+        callResults.map(e => ({
+          orderHash: e.orderHash,
+          price: e.price,
+          remaining: e.remaining,
+          category: e.category,
+          ...((!workerpool || workerpool === NULL_BYTES32) && {
+            workerpool: e.workerpool,
+          }),
+          ...(e.tag !== NULL_BYTES32 && { tag: e.tag }),
+          ...(e.trust > 1 && { trust: e.trust }),
+          ...(e.apprestrict !== NULL_ADDRESS && {
+            apprestrict: e.apprestrict,
+          }),
+          ...(e.datasetrestrict !== NULL_ADDRESS && {
+            datasetrestrict: e.datasetrestrict,
+          }),
+          ...(e.requesterrestrict !== NULL_ADDRESS && {
+            requesterrestrict: e.requesterrestrict,
+          }),
+        })),
+      )}`;
 
-      const successMessage = workerpoolOrders.length > 0
-        ? `Orderbook\nWorkerpool orders details:${pretty(workerpoolOrders)}\n`
-        : 'Empty orderbook';
+      const { results, count } = await displayPaginableRequest({
+        request,
+        processResponse,
+        fetchMessage,
+        createResultsMessage,
+        spinner,
+        raw,
+      });
+
+      const successMessage = results.length > 0 ? 'No more results' : 'Empty orderbook';
 
       spinner.succeed(successMessage, {
         raw: {
-          workerpoolOrders: response.workerpoolOrders,
+          count,
+          workerpoolOrders: results,
         },
       });
       spinner.info('Trade in the browser at https://market.iex.ec');
@@ -247,38 +367,44 @@ orderbookRequester
       const chain = await loadChain(cmd.chain, {
         spinner,
       });
-      spinner.start(info.showing(objName));
       const {
         tag,
-        maxTag,
         requireTag,
+        maxTag,
         app,
         dataset,
         workerpool,
         beneficiary,
+        category,
+        maxTrust,
+        minVolume,
+        raw,
       } = cmd;
-      const response = await orderbook.fetchRequestOrderbook(
+
+      const request = orderbook.fetchRequestOrderbook(
         chain.contracts,
         getPropertyFormChain(chain, 'iexecGateway'),
         {
-          category: cmd.category,
+          category,
           requester: address,
           minTag: tag !== undefined ? tag : requireTag,
           maxTag: tag !== undefined ? tag : maxTag,
-          minVolume: cmd.minVolume,
-          maxTrust: cmd.maxTrust,
+          minVolume,
+          maxTrust,
           app,
           dataset,
           workerpool,
           beneficiary,
         },
       );
-      const requestOrders = response.requestOrders
-        ? response.requestOrders.map(e => ({
+      const fetchMessage = info.showing(objName);
+      const processResponse = res => (res.orders
+        ? res.orders.map(e => ({
           orderHash: e.orderHash,
-          requester: e.order.requester,
           app: e.order.app,
           dataset: e.order.dataset,
+          workerpool: e.order.workerpool,
+          requester: e.order.requester,
           beneficiary: e.order.beneficiary,
           category: e.order.category,
           tag: e.order.tag,
@@ -286,15 +412,47 @@ orderbookRequester
           price: e.order.workerpoolmaxprice,
           remaining: e.remaining,
         }))
-        : [];
+        : []);
+      const createResultsMessage = (
+        callResults,
+        initilResultsCount,
+        totalCount,
+      ) => `Requestorders details (${initilResultsCount
+          + 1} to ${initilResultsCount + callResults.length}${
+        totalCount ? ` of ${totalCount}` : ''
+      }):${pretty(
+        callResults.map(e => ({
+          orderHash: e.orderHash,
+          price: e.price,
+          remaining: e.remaining,
+          category: e.category,
+          app: e.app,
+          ...(e.dataset !== NULL_ADDRESS && {
+            dataset: e.dataset,
+          }),
+          ...(e.tag !== NULL_BYTES32 && { tag: e.tag }),
+          ...(e.trust > 1 && { trust: e.trust }),
+          ...(e.workerpool !== NULL_ADDRESS && {
+            workerpool: e.workerpool,
+          }),
+        })),
+      )}`;
 
-      const successMessage = requestOrders.length > 0
-        ? `Orderbook\nRequest orders details:${pretty(requestOrders)}\n`
-        : 'Empty orderbook';
+      const { results, count } = await displayPaginableRequest({
+        request,
+        processResponse,
+        fetchMessage,
+        createResultsMessage,
+        spinner,
+        raw,
+      });
+
+      const successMessage = results.length > 0 ? 'No more results' : 'Empty orderbook';
 
       spinner.succeed(successMessage, {
         raw: {
-          requestOrders: response.requestOrders,
+          count,
+          requestOrders: results,
         },
       });
       spinner.info('Trade in the browser at https://market.iex.ec');